% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TEMPTED.R
\name{bernoulli_kernel}
\alias{bernoulli_kernel}
\title{Caculate the Bernoulli kernel}
\usage{
bernoulli_kernel(x, y)
}
\arguments{
\item{x, y}{Two values between which the Bernoulli kernel is calculated.}
}
\value{
The calculated kernel between \code{x} and \code{y}.
}
\description{
This function is used to calculate the kernel matrix for the
RKHS regression that iteratively updates the temporal loading function.
}
