% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TEMPTED.R
\docType{data}
\name{count_table}
\alias{count_table}
\title{OTU read count table from the ECAM data}
\format{
A data.frame with rows representing samples and matching with data.frame \code{meta_table}
and columns representing microbial features (i.e. OTUs). Each entry is a read count.
}
\usage{
count_table
}
\description{
OTU read count table from the ECAM data
}
\references{
Bokulich, Nicholas A., et al. "Antibiotics, birth mode, and diet shape microbiome maturation during early life." Science translational medicine 8.343 (2016): 343ra82-343ra82.
}
\keyword{datasets}
