% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TEMPTED.R
\name{est_test_subject}
\alias{est_test_subject}
\title{Estimate subject loading of testing data}
\usage{
est_test_subject(datlist, res_tempted, mean_svd = NULL)
}
\arguments{
\item{datlist}{Testing data formatted into datlist in the same fashion as the training data.
The same transformation needs to be used for both training and testing data.}

\item{res_tempted}{Result from \code{\link{tempted}} ran on the training data.}

\item{mean_svd}{Result from \code{\link{svd_centralize}} ran on the training data.}
}
\value{
estimated subject loading of testing data
}
\description{
This function estimates the subject loading of the testing data
based on feature and temporal loading from training data,
so that both the testing data and training data have the same dimensionality reduction.
}
\examples{
# split the example data into training and testing

id_test <- meta_table$studyid=="1"

count_train <- count_table[!id_test,]
meta_train <- meta_table[!id_test,]

count_test <- count_table[id_test,]
meta_test <- meta_table[id_test,]

# run tempted on training data

datlist_train <- format_tempted(count_train,
                                meta_train$day_of_life,
                                meta_train$studyid,
                                threshold=0.95,
                                pseudo_count=0.5,
                                transform="clr")

mean_svd_train <- svd_centralize(datlist_train, r=1)

res_tempted_train <- tempted(mean_svd_train$datlist,
r=3, smooth=1e-5)

# get the overlapping features

count_test <- count_test[,rownames(datlist_train[[1]])[-1]]

datlist_test <- format_tempted(count_test,
                               meta_test$day_of_life,
                               meta_test$studyid,
                               threshold=1,
                               pseudo_count=0.5,
                               transform="clr")

# estimate the subject loading of the testing subject

sub_test <- est_test_subject(datlist_test, res_tempted_train, mean_svd_train)
}
