% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TEMPTED.R
\name{freg_rkhs}
\alias{freg_rkhs}
\title{RKHS regression to update the temporal loading function.}
\usage{
freg_rkhs(Ly, a.hat, ind_vec, Kmat, Kmat_output, smooth = 1e-08)
}
\description{
This function is an internal function that performs the RKHS regression.
It is applied iteratively in \code{\link{tempted}} to update the temporal loading function.
}
