% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TEMPTED.R
\docType{data}
\name{meta_table}
\alias{meta_table}
\title{Meta data table from the ECAM data}
\format{
A data.frame with rows representing samples and matching with data.frame \code{count_table} and \code{processed_table}
and three columns:
\describe{
\item{studyid}{character denoting the subject ID of the infants.}
\item{delivery}{character denoting the delivery mode of the infants.}
\item{day_of_life}{character denoting the age of infants measured in days when microbiome sample was taken.}
}
}
\usage{
meta_table
}
\description{
Meta data table from the ECAM data
}
\references{
Bokulich, Nicholas A., et al. "Antibiotics, birth mode, and diet shape microbiome maturation during early life." Science translational medicine 8.343 (2016): 343ra82-343ra82.
}
\keyword{datasets}
