% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TEMPTED.R
\name{plot_time_loading}
\alias{plot_time_loading}
\title{Plot the temporal loading functions}
\usage{
plot_time_loading(res, r = NULL, ...)
}
\arguments{
\item{res}{Output of function \code{\link{tempted}}.}

\item{r}{The number of components to plot. By default all the components estimated by \code{\link{tempted}} will be plotted.}

\item{...}{Arguments to put in \code{ggplot2::geom_line(aes(...))}.}
}
\value{
An ggplot2 object.
}
\description{
This function uses \code{ggplot2::geom_line()} in ggplot2 to plot the temporal loading functions from \code{\link{tempted}}.
}
\seealso{
Examples can be found in \code{\link{tempted_all}} and \code{\link{tempted}}.
}
