% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TEMPTED.R
\docType{data}
\name{processed_table}
\alias{processed_table}
\title{Central-log-ratio (clr) transformed OTU table from the ECAM data}
\format{
A data.frame with rows representing samples and matching with data.frame \code{meta_table}
and columns representing microbial features (i.e. OTUs).
Entries do not need to be transformed, and will be directly used by \code{\link{tempted}}.
This data.frame is used to illustrate how \code{\link{tempted}} can be used for
general form of multivariate longitudinal data already preprocessed by user.
}
\usage{
processed_table
}
\description{
Central-log-ratio (clr) transformed OTU table from the ECAM data
}
\references{
Bokulich, Nicholas A., et al. "Antibiotics, birth mode, and diet shape microbiome maturation during early life." Science translational medicine 8.343 (2016): 343ra82-343ra82.
}
\keyword{datasets}
