% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TEMPTED.R
\name{tdenoise}
\alias{tdenoise}
\title{Calculate the de-noised temporal tensor}
\usage{
tdenoise(res_tempted, mean_svd = NULL)
}
\arguments{
\item{res_tempted}{Output of tempted}

\item{mean_svd}{Output of svd_centralize}
}
\value{
The de-noised functional tensor
}
\description{
This function constructs a de-noised version of the temporal tensor
using the low-rank components obtained by \code{\link{svd_centralize}} \code{\link{tempted}} and uses the loadings to
}
